[Refactor] Split odla_dnnl.cc to small pieces
